verified pagination service is working as intended

+++ backend/src/main/java/com/me/projects/articles/service/PaginationService.java
@@ -4,6 +4,6 @@ import com.me.projects.articles.model.Paginate;
 
 public interface PaginationService<T> {
 
-  Paginate<T> getPagination(long page, int pageSize) throws PaginationServiceException;
+  Paginate<T> getPagination(int page, int pageSize) throws PaginationServiceException;
 
 }
